requirements.txt
fastapi
uvicorn
SQLAlchemy[asyncio]
asyncpg
python-jose[cryptography]
passlib[bcrypt]
aioredis
pydantic
fastapi-limiter
